refactor: PCHeader 분리
