+++ src/components/common/Header/Header.module.scss
@@ -387,100 +387,6 @@
   }
 }
 
-.megamenu {
-  &__panel {
-    display: block;
-    top: 80px;
-    overflow: hidden;
-    position: absolute;
-    left: 0;
-    margin: 0 auto;
-    width: 100%;
-    height: 0;
-    min-width: 1132px;
-    z-index: 5;
-    background: #15507a;
-    transition: height 0.5s;
-  }
-
-  &:hover,
-  &:focus-within {
-    & .megamenu__panel {
-      height: 75px;
-      box-sizing: border-box;
-    }
-  }
-
-  &__trigger-list {
-    list-style: none;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    flex: 1;
-    height: 100%;
-  }
-
-  &__trigger {
-    display: block;
-    border: none;
-    float: left;
-    padding: 0 20px;
-    height: 80px;
-    font-weight: 700;
-    font-size: 17px;
-    text-align: left;
-    font-family: NanumSquare, serif;
-    line-height: 1;
-    color: #fff;
-    cursor: pointer;
-    background-color: #175c8e;
-
-    & > span {
-      padding: 31px 0 30px;
-    }
-
-    &:hover,
-    &:focus {
-      & > span {
-        padding-bottom: 28px;
-        border-bottom: 2px solid #fff;
-      }
-    }
-  }
-
-  &__content {
-    display: flex;
-    overflow: hidden;
-    width: calc(1132px - 135px);
-    height: 100%;
-    z-index: 6;
-    padding-left: 135px;
-    margin: 0 auto;
-    transition: height 0.2s, -webkit-transform 0.5s;
-    transition: height 0.2s, transform 0.5s;
-  }
-
-  &__menu {
-    padding: 26px 0;
-    margin-right: 38px;
-  }
-
-  &__link {
-    line-height: 2;
-    display: block;
-    font-size: 13px;
-    padding-bottom: 10px;
-    color: rgb(255 255 255);
-    text-decoration: none;
-    text-align: left;
-    font-weight: 400;
-
-    &:hover {
-      font-weight: 800;
-    }
-  }
-}
-
 @keyframes slide-left-enter {
   0% {
     opacity: 0;
