+++ src/components/common/Header/PCHeader/PCHeader.module.scss
@@ -1,0 +1,155 @@
+.header {
+  &__logo {
+    padding: 18px 35px 18px 0;
+
+    & > img {
+      width: 80px;
+      height: 44px;
+    }
+  }
+
+  &__megamenu {
+    height: 100%;
+    margin-right: auto;
+  }
+
+  &__auth-menu {
+    display: flex;
+    align-items: center;
+  }
+
+  &__auth-link {
+    font-weight: normal;
+    float: right;
+    text-align: right;
+    text-decoration: none;
+    font-family: NanumSquare, serif;
+    line-height: 1.35;
+    font-size: 15px;
+    color: #a0d2f6;
+    padding: 0 20px;
+    margin: 30px 0;
+    cursor: pointer;
+
+    &:hover {
+      font-weight: 800;
+      color: #fff;
+    }
+
+    & > a {
+      color: inherit;
+      text-decoration: none;
+    }
+  }
+
+  &__auth-button {
+    font-weight: normal;
+    background: none;
+    border: none;
+    color: #a0d2f6;
+    text-decoration: none;
+    font-family: NanumSquare, serif;
+    line-height: 1.35;
+    font-size: 15px;
+    cursor: pointer;
+
+    &:hover {
+      font-weight: 800;
+      color: #fff;
+    }
+  }
+}
+
+.megamenu {
+  &__trigger-list {
+    list-style: none;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    flex: 1;
+    height: 100%;
+  }
+
+  &__trigger {
+    display: block;
+    border: none;
+    float: left;
+    padding: 0 20px;
+    height: 80px;
+    font-weight: 700;
+    font-size: 17px;
+    text-align: left;
+    font-family: NanumSquare, serif;
+    line-height: 1;
+    color: #fff;
+    cursor: pointer;
+    background-color: #175c8e;
+
+    & > span {
+      padding: 31px 0 30px;
+    }
+
+    &:hover,
+    &:focus {
+      & > span {
+        padding-bottom: 28px;
+        border-bottom: 2px solid #fff;
+      }
+    }
+  }
+
+  &__panel {
+    display: block;
+    overflow: hidden;
+    position: absolute;
+    top: 80px;
+    left: 0;
+    margin: 0 auto;
+    width: 100%;
+    height: 0;
+    min-width: 1132px;
+    z-index: 5;
+    background: #15507a;
+    transition: height 0.5s;
+  }
+
+  &:hover,
+  &:focus-within {
+    & .megamenu__panel {
+      height: 75px;
+      box-sizing: border-box;
+    }
+  }
+
+  &__content {
+    display: flex;
+    overflow: hidden;
+    width: calc(1132px - 135px);
+    height: 100%;
+    z-index: 6;
+    padding-left: 135px;
+    margin: 0 auto;
+    transition: height 0.2s, -webkit-transform 0.5s;
+    transition: height 0.2s, transform 0.5s;
+  }
+
+  &__menu {
+    padding: 26px 0;
+    margin-right: 38px;
+  }
+
+  &__link {
+    line-height: 2;
+    display: block;
+    font-size: 13px;
+    padding-bottom: 10px;
+    color: rgb(255 255 255);
+    text-decoration: none;
+    text-align: left;
+    font-weight: 400;
+
+    &:hover {
+      font-weight: 800;
+    }
+  }
+}
